New blocked step interaction...
I'm not the pop-up adds to "demo story" here. Also the layout was kind of wonky. For simplicity sake, I just indicated it might a point to connect with help, though it's something we'll probably talk through about the different options here. Also removes the pop-up, which felt kind of weird and "system-y."